changed the mapping of duty cycle in PWM mode from 0 --> 100 to 0---> 1000

+++ AVR System/MCAL/MTIMER/MTIMER_interface.h
@@ -73,7 +73,7 @@ void MTIMER_voidSetOVCallback(u8 Copy_u8Timer, void (*func)(void));
 void MTIMER_voidSetCompareMatchCallback(u8 Copy_u8Timer, void (*func)(void));
 
 
-void MTIMER_voidSetPwmWithFreq(u32 Copy_u32Preiod_ms, u8 Copy_u8DutyCycle);
+void MTIMER_voidSetPwmWithFreq(u32 Copy_u32Preiod_ms, u16 Copy_u8DutyCycle);
 
 /**
  * Starts timer
